(common) Move AABB collision functions to common/aabb.ts
