+++ RigidBody/systems/sys_collide.ts
@@ -1,9 +1,7 @@
-import {get_translation} from "../../common/mat4.js";
-import {Vec3} from "../../common/math.js";
-import {negate, transform_point} from "../../common/vec3.js";
+import {compute_aabb, intersect_aabb, penetrate_aabb} from "../../common/aabb.js";
+import {negate} from "../../common/vec3.js";
 import {Collide} from "../components/com_collide.js";
 import {Has} from "../components/com_index.js";
-import {Transform} from "../components/com_transform.js";
 import {Game} from "../game.js";
 
 const QUERY = Has.Transform | Has.Collide;
@@ -21,9 +19,9 @@ export function sys_collide(game: Game, delta: number) {
             collider.Collisions = [];
             if (collider.New) {
                 collider.New = false;
-                compute_aabb(transform, collider);
+                compute_aabb(transform.World, collider);
             } else if (collider.Dynamic) {
-                compute_aabb(transform, collider);
+                compute_aabb(transform.World, collider);
                 dynamic_colliders.push(collider);
             } else {
                 static_colliders.push(collider);
@@ -65,96 +63,4 @@ function check_collisions(collider: Collide, colliders: Collide[], length: numbe
             });
         }
     }
-}
-
-const BOX = [
-    [0.5, 0.5, 0.5],
-    [0.5, 0.5, -0.5],
-    [-0.5, 0.5, -0.5],
-    [-0.5, 0.5, 0.5],
-    [0.5, -0.5, 0.5],
-    [0.5, -0.5, -0.5],
-    [-0.5, -0.5, -0.5],
-    [-0.5, -0.5, 0.5],
-];
-
-function compute_aabb(transform: Transform, collide: Collide) {
-    get_translation(collide.Center, transform.World);
-
-    // Start with the extents on each axis set to the position of the center.
-    let min_x, min_y, min_z, max_x, max_y, max_z;
-    min_x = max_x = collide.Center[0];
-    min_y = max_y = collide.Center[1];
-    min_z = max_z = collide.Center[2];
-
-    // Expand the extents outwards from the center by finding the farthest
-    // vertex on each axis in both the negative and the positive direction.
-    let world_vertex = <Vec3>[0, 0, 0];
-    for (let i = 0; i < 8; i++) {
-        let bb_vertex = BOX[i];
-
-        // Scale the bounding box according to the size of the collider.
-        world_vertex[0] = bb_vertex[0] * collide.Size[0];
-        world_vertex[1] = bb_vertex[1] * collide.Size[1];
-        world_vertex[2] = bb_vertex[2] * collide.Size[2];
-
-        transform_point(world_vertex, world_vertex, transform.World);
-        if (world_vertex[0] < min_x) {
-            min_x = world_vertex[0];
-        }
-        if (world_vertex[0] > max_x) {
-            max_x = world_vertex[0];
-        }
-        if (world_vertex[1] < min_y) {
-            min_y = world_vertex[1];
-        }
-        if (world_vertex[1] > max_y) {
-            max_y = world_vertex[1];
-        }
-        if (world_vertex[2] < min_z) {
-            min_z = world_vertex[2];
-        }
-        if (world_vertex[2] > max_z) {
-            max_z = world_vertex[2];
-        }
-    }
-
-    // Save the min and max bounds.
-    collide.Min = [min_x, min_y, min_z];
-    collide.Max = [max_x, max_y, max_z];
-
-    // Calculate the half-extents.
-    collide.Half[0] = (max_x - min_x) / 2;
-    collide.Half[1] = (max_y - min_y) / 2;
-    collide.Half[2] = (max_z - min_z) / 2;
-}
-
-function penetrate_aabb(a: Collide, b: Collide) {
-    let distance_x = a.Center[0] - b.Center[0];
-    let penetration_x = a.Half[0] + b.Half[0] - Math.abs(distance_x);
-
-    let distance_y = a.Center[1] - b.Center[1];
-    let penetration_y = a.Half[1] + b.Half[1] - Math.abs(distance_y);
-
-    let distance_z = a.Center[2] - b.Center[2];
-    let penetration_z = a.Half[2] + b.Half[2] - Math.abs(distance_z);
-
-    if (penetration_x < penetration_y && penetration_x < penetration_z) {
-        return <Vec3>[penetration_x * Math.sign(distance_x), 0, 0];
-    } else if (penetration_y < penetration_z) {
-        return <Vec3>[0, penetration_y * Math.sign(distance_y), 0];
-    } else {
-        return <Vec3>[0, 0, penetration_z * Math.sign(distance_z)];
-    }
-}
-
-function intersect_aabb(a: Collide, b: Collide) {
-    return (
-        a.Min[0] < b.Max[0] &&
-        a.Max[0] > b.Min[0] &&
-        a.Min[1] < b.Max[1] &&
-        a.Max[1] > b.Min[1] &&
-        a.Min[2] < b.Max[2] &&
-        a.Max[2] > b.Min[2]
-    );
 }
